Fix fas credentialstore for demo (#72)
* Fix for FAS and Credential Store

* fixes

* fix fetch identity and aws creds actions
- cleanup plugin FAS api call
- fix listener cancellations

* Cleaned up FAS in the cognito plugin

+++ aws-auth-cognito/src/main/java/com/amplifyframework/auth/cognito/data/AmplifyCredential.kt
@@ -14,7 +14,7 @@ data class CognitoUserPoolTokens(
     val idToken: String?,
     val accessToken: String?,
     val refreshToken: String?,
-    val tokenExpiration: Int?,
+    val tokenExpiration: Long?,
 )
 
 @Serializable
